Ensure the description max length is also set in the initial form creation
This helps avoid a bug ultimately arising in ultimateForm whereby a formSave operation doesn't error on an over-long result.

So, for instance, if the user created an initial title of 167 chars, this will pass the initial form, but then when a formSave is done in the main form, this fails validation but is not reported. The form then returns an empty value because a problem has been found for that field.